[#625] client/container: remove intermediate wrapper
