+++ pkg/innerring/settlement.go
@@ -16,7 +16,7 @@ import (
 	"github.com/nspcc-dev/neofs-node/pkg/innerring/processors/settlement/common"
 	auditClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/audit"
 	balanceClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/balance"
-	containerClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/container/wrapper"
+	containerClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/container"
 	"github.com/nspcc-dev/neofs-node/pkg/util/logger"
 	auditAPI "github.com/nspcc-dev/neofs-sdk-go/audit"
 	containerAPI "github.com/nspcc-dev/neofs-sdk-go/container"
@@ -55,7 +55,7 @@ type auditSettlementDeps struct {
 
 type basicIncomeSettlementDeps struct {
 	*settlementDeps
-	cnrClient *containerClient.Wrapper
+	cnrClient *containerClient.Client
 }
 
 type basicSettlementConstructor struct {
